content/canvas/test/test_2d.path.arc.scale.1.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.path.arc.scale.1</title>
<!-- Testing: Non-uniformly scaled arcs are the right shape -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
function isPixel(ctx, x,y, r,g,b,a, pos, colour, d) {
    var pixel = ctx.getImageData(x, y, 1, 1);
    var pr = pixel.data[0],
        pg = pixel.data[1],
        pb = pixel.data[2],
        pa = pixel.data[3];
    ok(r-d <= pr && pr <= r+d &&
       g-d <= pg && pg <= g+d &&
       b-d <= pb && pb <= b+d &&
       a-d <= pa && pa <= a+d,
       "pixel "+pos+" is "+pr+","+pg+","+pb+","+pa+"; expected "+colour+" +/- "+d);
}

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
ctx.scale(2, 0.5);
ctx.fillStyle = '#0f0';
ctx.beginPath();
ctx.arc(25, 50, 56, 0, 2*Math.PI, false);
ctx.fill();
ctx.fillStyle = '#f00';
ctx.beginPath();
ctx.moveTo(-25, 50);
ctx.arc(-25, 50, 24, 0, 2*Math.PI, false);
ctx.moveTo(75, 50);
ctx.arc(75, 50, 24, 0, 2*Math.PI, false);
ctx.moveTo(25, -25);
ctx.arc(25, -25, 24, 0, 2*Math.PI, false);
ctx.moveTo(25, 125);
ctx.arc(25, 125, 24, 0, 2*Math.PI, false);
ctx.fill();

isPixel(ctx, 0,0, 0,255,0,255, "0,0", "0,255,0,255", 0);
isPixel(ctx, 50,0, 0,255,0,255, "50,0", "0,255,0,255", 0);
isPixel(ctx, 99,0, 0,255,0,255, "99,0", "0,255,0,255", 0);
isPixel(ctx, 0,25, 0,255,0,255, "0,25", "0,255,0,255", 0);
isPixel(ctx, 50,25, 0,255,0,255, "50,25", "0,255,0,255", 0);
isPixel(ctx, 99,25, 0,255,0,255, "99,25", "0,255,0,255", 0);
isPixel(ctx, 0,49, 0,255,0,255, "0,49", "0,255,0,255", 0);
isPixel(ctx, 50,49, 0,255,0,255, "50,49", "0,255,0,255", 0);
isPixel(ctx, 99,49, 0,255,0,255, "99,49", "0,255,0,255", 0);

SimpleTest.finish();

});
</script>
